feat(frontend): reorg insert converters and introduce stmt_to_region
