+++ src/api/src/helper.rs
@@ -810,50 +810,54 @@ pub fn vectors_to_rows<'a>(
     let mut rows = vec![Row { values: vec![] }; row_count];
     for column in columns {
         for (row_index, row) in rows.iter_mut().enumerate() {
-            row.values.push(GrpcValue {
-                value_data: match column.get(row_index) {
-                    Value::Null => None,
-                    Value::Boolean(v) => Some(ValueData::BoolValue(v)),
-                    Value::UInt8(v) => Some(ValueData::U8Value(v as _)),
-                    Value::UInt16(v) => Some(ValueData::U16Value(v as _)),
-                    Value::UInt32(v) => Some(ValueData::U32Value(v)),
-                    Value::UInt64(v) => Some(ValueData::U64Value(v)),
-                    Value::Int8(v) => Some(ValueData::I8Value(v as _)),
-                    Value::Int16(v) => Some(ValueData::I16Value(v as _)),
-                    Value::Int32(v) => Some(ValueData::I32Value(v)),
-                    Value::Int64(v) => Some(ValueData::I64Value(v)),
-                    Value::Float32(v) => Some(ValueData::F32Value(*v)),
-                    Value::Float64(v) => Some(ValueData::F64Value(*v)),
-                    Value::String(v) => Some(ValueData::StringValue(v.as_utf8().to_string())),
-                    Value::Binary(v) => Some(ValueData::BinaryValue(v.to_vec())),
-                    Value::Date(v) => Some(ValueData::DateValue(v.val())),
-                    Value::DateTime(v) => Some(ValueData::DatetimeValue(v.val())),
-                    Value::Timestamp(v) => Some(match v.unit() {
-                        TimeUnit::Second => ValueData::TimeSecondValue(v.value()),
-                        TimeUnit::Millisecond => ValueData::TimeMillisecondValue(v.value()),
-                        TimeUnit::Microsecond => ValueData::TimeMicrosecondValue(v.value()),
-                        TimeUnit::Nanosecond => ValueData::TimeNanosecondValue(v.value()),
-                    }),
-                    Value::Time(v) => Some(match v.unit() {
-                        TimeUnit::Second => ValueData::TimeSecondValue(v.value()),
-                        TimeUnit::Millisecond => ValueData::TimeMillisecondValue(v.value()),
-                        TimeUnit::Microsecond => ValueData::TimeMicrosecondValue(v.value()),
-                        TimeUnit::Nanosecond => ValueData::TimeNanosecondValue(v.value()),
-                    }),
-                    Value::Interval(v) => Some(match v.unit() {
-                        IntervalUnit::YearMonth => ValueData::IntervalYearMonthValues(v.to_i32()),
-                        IntervalUnit::DayTime => ValueData::IntervalDayTimeValues(v.to_i64()),
-                        IntervalUnit::MonthDayNano => ValueData::IntervalMonthDayNanoValues(
-                            convert_i128_to_interval(v.to_i128()),
-                        ),
-                    }),
-                    Value::List(_) => unreachable!(),
-                },
-            })
+            row.values.push(value_to_grpc_value(column.get(row_index)))
         }
     }
 
     rows
+}
+
+pub fn value_to_grpc_value(value: Value) -> GrpcValue {
+    GrpcValue {
+        value_data: match value {
+            Value::Null => None,
+            Value::Boolean(v) => Some(ValueData::BoolValue(v)),
+            Value::UInt8(v) => Some(ValueData::U8Value(v as _)),
+            Value::UInt16(v) => Some(ValueData::U16Value(v as _)),
+            Value::UInt32(v) => Some(ValueData::U32Value(v)),
+            Value::UInt64(v) => Some(ValueData::U64Value(v)),
+            Value::Int8(v) => Some(ValueData::I8Value(v as _)),
+            Value::Int16(v) => Some(ValueData::I16Value(v as _)),
+            Value::Int32(v) => Some(ValueData::I32Value(v)),
+            Value::Int64(v) => Some(ValueData::I64Value(v)),
+            Value::Float32(v) => Some(ValueData::F32Value(*v)),
+            Value::Float64(v) => Some(ValueData::F64Value(*v)),
+            Value::String(v) => Some(ValueData::StringValue(v.as_utf8().to_string())),
+            Value::Binary(v) => Some(ValueData::BinaryValue(v.to_vec())),
+            Value::Date(v) => Some(ValueData::DateValue(v.val())),
+            Value::DateTime(v) => Some(ValueData::DatetimeValue(v.val())),
+            Value::Timestamp(v) => Some(match v.unit() {
+                TimeUnit::Second => ValueData::TimeSecondValue(v.value()),
+                TimeUnit::Millisecond => ValueData::TimeMillisecondValue(v.value()),
+                TimeUnit::Microsecond => ValueData::TimeMicrosecondValue(v.value()),
+                TimeUnit::Nanosecond => ValueData::TimeNanosecondValue(v.value()),
+            }),
+            Value::Time(v) => Some(match v.unit() {
+                TimeUnit::Second => ValueData::TimeSecondValue(v.value()),
+                TimeUnit::Millisecond => ValueData::TimeMillisecondValue(v.value()),
+                TimeUnit::Microsecond => ValueData::TimeMicrosecondValue(v.value()),
+                TimeUnit::Nanosecond => ValueData::TimeNanosecondValue(v.value()),
+            }),
+            Value::Interval(v) => Some(match v.unit() {
+                IntervalUnit::YearMonth => ValueData::IntervalYearMonthValues(v.to_i32()),
+                IntervalUnit::DayTime => ValueData::IntervalDayTimeValues(v.to_i64()),
+                IntervalUnit::MonthDayNano => {
+                    ValueData::IntervalMonthDayNanoValues(convert_i128_to_interval(v.to_i128()))
+                }
+            }),
+            Value::List(_) => unreachable!(),
+        },
+    }
 }
 
 /// Returns true if the column type is equal to expected type.
