+++ src/frontend/src/inserter/req_convert.rs
@@ -1,0 +1,61 @@
+// Copyright 2023 Greptime Team
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+pub mod column_to_row;
+pub mod row_to_region;
+pub mod stmt_to_region;
+pub mod table_to_region;
+
+use api::helper::ColumnDataTypeWrapper;
+use api::v1::{ColumnDataType, SemanticType};
+use datatypes::prelude::ConcreteDataType;
+use snafu::{OptionExt, ResultExt};
+use table::metadata::TableInfo;
+
+use crate::error::{ColumnDataTypeSnafu, ColumnNotFoundSnafu, MissingTimeIndexColumnSnafu, Result};
+
+fn semantic_type(table_info: &TableInfo, column: &str) -> Result<SemanticType> {
+    let table_meta = &table_info.meta;
+    let table_schema = &table_meta.schema;
+
+    let time_index_column = &table_schema
+        .timestamp_column()
+        .with_context(|| table::error::MissingTimeIndexColumnSnafu {
+            table_name: table_info.name.to_string(),
+        })
+        .context(MissingTimeIndexColumnSnafu)?
+        .name;
+
+    let semantic_type = if column == time_index_column {
+        SemanticType::Timestamp
+    } else {
+        let column_index = table_schema.column_index_by_name(column);
+        let column_index = column_index.context(ColumnNotFoundSnafu {
+            msg: format!("unable to find column {column} in table schema"),
+        })?;
+
+        if table_meta.primary_key_indices.contains(&column_index) {
+            SemanticType::Tag
+        } else {
+            SemanticType::Field
+        }
+    };
+
+    Ok(semantic_type)
+}
+
+fn data_type(data_type: ConcreteDataType) -> Result<ColumnDataType> {
+    let datatype: ColumnDataTypeWrapper = data_type.try_into().context(ColumnDataTypeSnafu)?;
+    Ok(datatype.datatype())
+}
