+++ src/frontend/src/instance/distributed.rs
@@ -39,7 +39,6 @@ use common_meta::table_name::TableName;
 use common_query::Output;
 use common_telemetry::info;
 use datanode::instance::sql::table_idents_to_full_name;
-use datanode::sql::SqlHandler;
 use datatypes::prelude::ConcreteDataType;
 use datatypes::schema::RawSchema;
 use partition::manager::PartitionInfo;
@@ -61,11 +60,12 @@ use super::region_handler::RegionRequestHandler;
 use crate::catalog::FrontendCatalogManager;
 use crate::error::{
     self, AlterExprToRequestSnafu, CatalogSnafu, ColumnDataTypeSnafu, ColumnNotFoundSnafu,
-    DeserializePartitionSnafu, InvokeDatanodeSnafu, NotSupportedSnafu, ParseSqlSnafu, Result,
-    SchemaExistsSnafu, SchemaNotFoundSnafu, TableAlreadyExistSnafu, TableMetadataManagerSnafu,
-    TableNotFoundSnafu, UnrecognizedTableOptionSnafu,
+    DeserializePartitionSnafu, NotSupportedSnafu, ParseSqlSnafu, Result, SchemaExistsSnafu,
+    SchemaNotFoundSnafu, TableAlreadyExistSnafu, TableMetadataManagerSnafu, TableNotFoundSnafu,
+    UnrecognizedTableOptionSnafu,
 };
 use crate::expr_factory;
+use crate::inserter::req_convert::stmt_to_region::StatementToRegion;
 use crate::instance::distributed::deleter::DistDeleter;
 use crate::instance::distributed::inserter::DistInserter;
 use crate::table::DistTable;
@@ -271,13 +271,11 @@ impl DistInstance {
                 self.drop_table(table_name).await
             }
             Statement::Insert(insert) => {
-                let insert_request =
-                    SqlHandler::insert_to_request(self.catalog_manager.clone(), &insert, query_ctx)
-                        .await
-                        .context(InvokeDatanodeSnafu)?;
-
+                let request = StatementToRegion::new(self.catalog_manager.as_ref(), &query_ctx)
+                    .convert(&insert)
+                    .await?;
                 let inserter = DistInserter::new(&self.catalog_manager);
-                let affected_rows = inserter.insert_table_request(insert_request).await?;
+                let affected_rows = inserter.insert(request).await?;
                 Ok(Output::AffectedRows(affected_rows as usize))
             }
             Statement::ShowCreateTable(show) => {
@@ -612,7 +610,7 @@ impl RegionRequestHandler for DistRegionRequestHandler {
             region_request::Body::Inserts(inserts) => {
                 let inserter =
                     DistInserter::new(&self.catalog_manager).with_trace_id(ctx.trace_id());
-                let affected_rows = inserter.insert_region_requests(inserts).await? as _;
+                let affected_rows = inserter.insert(inserts).await? as _;
                 Ok(RegionResponse {
                     header: Some(Default::default()),
                     affected_rows,
